+++ docs/changelog.rst
@@ -2,6 +2,15 @@
 Changelog
 *********
 
+v2.1.3 (UNRELEASED)
+===================
+
+Maintenance release.
+
+- Document that the playlists API is broken. If it is used, emit a warning to notify
+  the user of the playlist functionality.
+
+
 v2.1.2 (2019-12-10)
 ===================